split out multistructure tests
