+++ tests/test_collections.py
@@ -14,6 +14,8 @@ from geostructures import GeoBox, GeoCircle, GeoLineString, GeoPoint, GeoPolygon
 from geostructures.time import TimeInterval
 from geostructures.collections import Track, FeatureCollection
 
+from tests.functions import pyshp_round_trip
+
 
 def test_collection_bool():
     track1 = Track([
@@ -460,7 +462,7 @@ def test_collection_convex_hull():
     ])
 
 
-def test_collection_to_from_shapefile(caplog):
+def test_collection_to_from_shapefile(pyshp_round_trip, caplog):
     # Tests both to and from because of temporary file usage
     # Shapes will be assigned an ID when written to shapefile, so have to hard code in tests
 
@@ -497,12 +499,8 @@ def test_collection_to_from_shapefile(caplog):
         )
     ])
 
-    with tempfile.TemporaryDirectory() as f:
-        with ZipFile(os.path.join(f, 'test.zip'), 'w') as zfile:
-            shapecol.to_shapefile(zfile)
-
-        new_shapecol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'))
-        assert set(new_shapecol.geoshapes) == set(shapecol.geoshapes)
+    new = pyshp_round_trip(shapecol)
+    assert set(new.geoshapes) == set(shapecol.geoshapes)
 
     # Test again with Z and M values
     shapecol = FeatureCollection([
@@ -524,6 +522,10 @@ def test_collection_to_from_shapefile(caplog):
             GeoLineString([Coordinate(0., 1., 99., 100.), Coordinate(1., 0., 99., 100.)]),
             GeoLineString([Coordinate(0., 0., 99., 100.), Coordinate(1., 1., 99., 100.)]),
         ], properties={'ID': 2}),
+        MultiGeoLineString([
+            GeoLineString([Coordinate(0., 1., m=1), Coordinate(1., 0., m=1)]),
+            GeoLineString([Coordinate(0., 0., m=2), Coordinate(1., 1., m=2)]),
+        ], properties={'ID': 2}),
         MultiGeoPolygon(
             [
                 GeoBox(
@@ -539,31 +541,25 @@ def test_collection_to_from_shapefile(caplog):
         )
     ])
 
-    with tempfile.TemporaryDirectory() as f:
-        with ZipFile(os.path.join(f, 'test.zip'), 'w') as zfile:
-            shapecol.to_shapefile(zfile)
+    new = pyshp_round_trip(shapecol)
+    assert set(new.geoshapes) == set(shapecol.geoshapes)
 
-        new_shapecol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'))
-        for x in new_shapecol:
-            if x not in shapecol:
-                raise
-        # assert set(new_shapecol.geoshapes) == set(shapecol.geoshapes)
-
-    # Test that non-polygons get written/read correctly (should be read as a polygon)
-    shapecol = FeatureCollection([
-        GeoBox(Coordinate(0.0, 1.0), Coordinate(1.0, 0.0), properties={'ID': 0}),
-        GeoCircle(Coordinate(0.0, 2.0), 1000, properties={'ID': 1}),
-        GeoRing(Coordinate(0.0, 2.0), 1000, 500, properties={'ID': 2}),
-    ])
-    with tempfile.TemporaryDirectory() as f:
-        with ZipFile(os.path.join(f, 'test.zip'), 'w') as zfile:
-            shapecol.to_shapefile(zfile)
-
-        new_shapecol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'))
-        assert set(new_shapecol.geoshapes) == set(x.to_polygon() for x in shapecol.geoshapes)
-
-        new_shapecol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'), read_layers=['nonexistent'])
-        assert new_shapecol.geoshapes == []
+    #
+    # # Test that non-polygons get written/read correctly (should be read as a polygon)
+    # shapecol = FeatureCollection([
+    #     GeoBox(Coordinate(0.0, 1.0), Coordinate(1.0, 0.0), properties={'ID': 0}),
+    #     GeoCircle(Coordinate(0.0, 2.0), 1000, properties={'ID': 1}),
+    #     GeoRing(Coordinate(0.0, 2.0), 1000, 500, properties={'ID': 2}),
+    # ])
+    # with tempfile.TemporaryDirectory() as f:
+    #     with ZipFile(os.path.join(f, 'test.zip'), 'w') as zfile:
+    #         shapecol.to_shapefile(zfile)
+    #
+    #     new_shapecol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'))
+    #     assert set(new_shapecol.geoshapes) == set(x.to_polygon() for x in shapecol.geoshapes)
+    #
+    #     new_shapecol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'), read_layers=['nonexistent'])
+    #     assert new_shapecol.geoshapes == []
 
     # Test writing/reading properties
     pointcol = FeatureCollection([
@@ -586,43 +582,41 @@ def test_collection_to_from_shapefile(caplog):
             }
         ),
     ])
-    with tempfile.TemporaryDirectory() as f:
-        with ZipFile(os.path.join(f, 'test.zip'), 'w') as zfile:
-            pointcol.to_shapefile(zfile)
+    new = pyshp_round_trip(pointcol)
+    assert set(new.geoshapes) == set(pointcol.geoshapes)
 
-        new_pointcol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'))
-        assert set(new_pointcol.geoshapes) == set(pointcol.geoshapes)
 
-    # Test limiting the properties written
-    with tempfile.TemporaryDirectory() as f:
-        with ZipFile(os.path.join(f, 'test.zip'), 'w') as zfile:
-            # should ignore the 'ex2' prop
-            pointcol.to_shapefile(zfile, include_properties=['ex_prop'])
-
-        new_pointcol = FeatureCollection.from_shapefile(os.path.join(f, 'test.zip'))
-
-        expected = FeatureCollection([
-            GeoPoint(
-                Coordinate(1.0, 0.0),
-                properties={'ID': 0, 'ex_prop': 'test2'}
-            ),
-            GeoPoint(
-                Coordinate(2.0, 0.0),
-                properties={'ID': 1, 'ex_prop': 'test'}
-            ),
-        ])
-        assert set(new_pointcol.geoshapes) == set(expected.geoshapes)
+    orig = FeatureCollection([
+        GeoPoint(
+            Coordinate(1.0, 0.0),
+            properties={'ID': 0, 'ex_prop': 'test2'}
+        ),
+        GeoPoint(
+            Coordinate(2.0, 0.0),
+            properties={'ID': 1, 'ex_prop': 'test'}
+        ),
+    ])
+    expected = FeatureCollection([
+        GeoPoint(
+            Coordinate(1.0, 0.0),
+            properties={'ex_prop': 'test2'}
+        ),
+        GeoPoint(
+            Coordinate(2.0, 0.0),
+            properties={'ex_prop': 'test'}
+        ),
+    ])
+    new = pyshp_round_trip(orig, include_properties=['ex_prop'])
+    assert set(new.geoshapes) == set(expected.geoshapes)
 
     # Test that writing mixed property data types logs a warning
-    with tempfile.TemporaryDirectory() as f:
-        pointcol = FeatureCollection([
-            GeoPoint(Coordinate(1.0, 0.0), properties={'ID': 0, 'prop': 1}),
-            GeoPoint(Coordinate(2.0, 0.0), properties={'ID': 1, 'prop': '2'}),
-        ])
-        with ZipFile(os.path.join(f, 'test.zip'), 'w') as zfile:
-            pointcol.to_shapefile(zfile)
-
-        assert 'Conflicting data types found in properties; your shapefile may not get written correctly' in caplog.text
+    caplog.clear()
+    pointcol = FeatureCollection([
+        GeoPoint(Coordinate(1.0, 0.0), properties={'ID': 0, 'prop': 1}),
+        GeoPoint(Coordinate(2.0, 0.0), properties={'ID': 1, 'prop': '2.0'}),
+    ])
+    pyshp_round_trip(pointcol)
+    assert 'Conflicting data types found in properties' in caplog.text
 
 
 def test_featurecollection_add():
